refactor: create Abi component
